% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{compare_locks}
\alias{compare_locks}
\title{Compare multiple lockfiles}
\usage{
compare_locks(x)
}
\arguments{
\item{x}{A list of lists or list of 'rt_tibble'. An output from \code{read_lock()}.}
}
\value{
A tibble
}
\description{
Compare multiple lockfiles
}
\examples{
paths <- c(
  file.path(system.file("extdata", package = "renvtools"), "renv-r4.4.1.lock"),
  file.path(system.file("extdata", package = "renvtools"), "renv-r4.3.2.lock")
)
l <- read_lock(paths, format = "list")
compare_locks(l)
}
