% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.r
\name{get_version_pkg}
\alias{get_version_pkg}
\title{Get package version}
\usage{
get_version_pkg(x, pkg = "renv")
}
\arguments{
\item{x}{A 'rt_list' or 'rt_tibble' class object. An single output from \code{read_lock()}.}

\item{pkg}{package name with correct case}
}
\value{
A character string with version or NA
}
\description{
Get package version from lockfile
}
\examples{
paths <- file.path(system.file("extdata", package = "renvtools"), "renv-r4.4.1.lock")
l <- read_lock(paths, format = "list")
get_version_pkg(l, "renv")
}
