% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{is.rt}
\alias{is.rt}
\title{Check if input is single item}
\usage{
is.rt(x)
}
\arguments{
\item{x}{An object of class rt_list or rt_tibble}
}
\value{
TRUE or FALSE
}
\description{
Check if input is a single rt_list or a single rt_tibble class
}
