% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{is.rt_list}
\alias{is.rt_list}
\title{Check if input is rt_list class}
\usage{
is.rt_list(x)
}
\arguments{
\item{x}{An object of class rt_list}
}
\value{
TRUE or FALSE
}
\description{
Check if input is rt_list class
}
