% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{is.rt_tibble}
\alias{is.rt_tibble}
\title{Check if an object is rt_tibble class}
\usage{
is.rt_tibble(x)
}
\arguments{
\item{x}{An object of class rt_tibble}
}
\value{
TRUE or FALSE
}
\description{
Check if an object is rt_tibble class
}
