% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{list_to_tbl}
\alias{list_to_tbl}
\title{Convert a list to a rt_tibble}
\usage{
list_to_tbl(x)
}
\arguments{
\item{x}{A list containing one lockfile. A list from \code{read_lock()}.}
}
\value{
A tibble of class 'rt_tibble'.
}
\description{
Convert a list to a rt_tibble
}
