% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{print.rt_list}
\alias{print.rt_list}
\title{Print function for rt_list}
\usage{
\method{print}{rt_list}(x, ...)
}
\arguments{
\item{x}{An object of class rt_list}

\item{...}{Other}
}
\description{
Print function for rt_list
}
\details{
Prints key features of the lockfile.
}
