% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\name{print.rt_tibble}
\alias{print.rt_tibble}
\title{Print function for rt_tibble}
\usage{
\method{print}{rt_tibble}(x, ...)
}
\arguments{
\item{x}{An object of class rt_tibble}

\item{...}{Other}
}
\description{
Print function for rt_tibble
}
\details{
Prints R version along with table view.
}
