% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{read_lock}
\alias{read_lock}
\title{Read lock files}
\usage{
read_lock(
  paths = "renv.lock",
  type = "list",
  method = "default",
  encoding = "UTF-8"
)
}
\arguments{
\item{paths}{Path to lock files}

\item{type}{Return object type. Either 'tibble' or 'list'}

\item{method}{Read method. Either 'default' or 'renv'. See details.}

\item{encoding}{Text encoding, See \code{jsonlite::fromJSON()}.}
}
\value{
A tibble or a list
}
\description{
Read lock files as tibble or list
}
\details{
If argument \code{paths} is longer than one, then a list is returned.
Argument \code{type} denotes whether to return a tibble or a list.
Argument \code{method} denotes which method to use to read a lockfile. 'default'
reads the lockfile and preserves it as it is as much as possible. 'renv' uses
the read function from renv. It might update the lockfile structure to be current.
}
\examples{
# read one lock file as a tibble
path <- file.path(system.file("extdata", package = "renvtools"), "renv-r4.4.1.lock")
read_lock(path, type = "tibble")

# read one lock file as a list
x <- read_lock(path, type = "list")

# read several lock files as a list of tibbles
paths <- list.files(system.file("extdata", package = "renvtools"), full.names = TRUE)
names(paths) <- basename(paths)
x <- read_lock(paths, type = "tibble")

# read several lock files as a list of lists
x <- read_lock(paths, type = "list")

}
