% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{tbl_to_list}
\alias{tbl_to_list}
\title{Convert a rt_tibble to a list}
\usage{
tbl_to_list(x)
}
\arguments{
\item{x}{A tibble of class 'rt_tibble' from one lockfile. A tibble from \code{read_lock()}.}
}
\value{
A list
}
\description{
Convert a rt_tibble to a list
}
