% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{write_lock}
\alias{write_lock}
\title{Export a lock file}
\usage{
write_lock(x, path)
}
\arguments{
\item{x}{A single list or a single 'rt_tibble'. An single file output from \code{read_lock()}.}

\item{path}{Output path with filename}
}
\value{
None
}
\description{
Write a rt_tibble or a list as a lock file
}
