% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{read_lock}
\alias{read_lock}
\title{Read lock files}
\usage{
read_lock(paths = "renv.lock", format = "list", encoding = "UTF-8")
}
\arguments{
\item{paths}{Path to lock files}

\item{format}{Return object format. Either 'list', 'tibble' or 'renv'. See details.}

\item{encoding}{Text encoding, See \code{jsonlite::fromJSON()}.}
}
\value{
A tibble or a list
}
\description{
Read lock files as tibble or list
}
\details{
If argument \code{paths} is longer than one, then a list is returned.
Argument \code{format} denotes return object format for the lockfile.
Argument \code{method} denotes which method to use to read a lockfile. 'list'
reads the lockfile and preserves it as it is as much as possible. 'tibble'
returns a tibble.'renv' uses the read function from renv
(\code{renv::lockfile_read()}) which returns a list.
This option might update the lockfile structure to be current.
}
\examples{
# read one lock file as a tibble
path <- file.path(system.file("extdata", package = "renvtools"), "renv-r4.4.1.lock")
read_lock(path, format = "tibble")

# read one lock file as a list
x <- read_lock(path, format = "list")

# read several lock files as a list of tibbles
paths <- list.files(system.file("extdata", package = "renvtools"), full.names = TRUE)
names(paths) <- basename(paths)
x <- read_lock(paths, format = "tibble")

# read several lock files as a list of lists
x <- read_lock(paths, format = "list")

}
