% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renvtools.r
\name{summarize_lock}
\alias{summarize_lock}
\title{Summarize lock files}
\usage{
summarize_lock(x)
}
\arguments{
\item{x}{A list of lists or list of 'rt_tibble'. An output from \code{read_lock()}.}
}
\value{
A tibble
}
\description{
Summarize lock files
}
\examples{
paths <- c(
  file.path(system.file("extdata", package = "renvtools"), "renv-r4.4.1.lock"),
  file.path(system.file("extdata", package = "renvtools"), "renv-r4.3.2.lock")
)
l <- read_lock(paths, format = "list")
summarize_lock(l)

}
